Cleaning meshing code
