+++ src/aliceVision/fuseCut/CMakeLists.txt
@@ -1,19 +1,28 @@
 # Headers
 set(fuseCut_files_headers
-  DelaunayGraphCut.hpp
   delaunayGraphCutTypes.hpp
   Fuser.hpp
   MaxFlow_AdjList.hpp
   Octree.hpp
   InputSet.hpp
+  BoundingBox.hpp
+  Intersections.hpp
+  PointCloud.hpp
+  GraphFiller.hpp
+  Mesher.hpp
 )
 
 # Sources
 set(fuseCut_files_sources
-  DelaunayGraphCut.cpp
+
   Fuser.cpp
   MaxFlow_AdjList.cpp
   InputSet.cpp
+  Tetrahedralization.cpp
+  Intersections.cpp
+  PointCloud.cpp
+  GraphFiller.cpp
+  Mesher.cpp
 )
 
 alicevision_add_library(aliceVision_fuseCut
