+++ src/aliceVision/fuseCut/DelaunayGraphCut_test.cpp
@@ -6,8 +6,9 @@
 
 #include <aliceVision/sfm/sfm.hpp>
 #include <aliceVision/multiview/NViewDataSet.hpp>
-#include <aliceVision/fuseCut/DelaunayGraphCut.hpp>
 #include <aliceVision/fuseCut/Fuser.hpp>
+#include <aliceVision/fuseCut/PointCloud.hpp>
+#include <aliceVision/fuseCut/GraphFiller.hpp>
 
 #include <boost/math/constants/constants.hpp>
 
@@ -68,7 +69,6 @@ BOOST_AUTO_TEST_CASE(fuseCut_delaunayGraphCut)
     mvsUtils::MultiViewParams mp(sfmData, "", "", "", false);
 
     mp.userParams.put("LargeScale.universePercentile", 0.999);
-    mp.userParams.put("delaunaycut.forceTEdgeDelta", 0.1f);
     mp.userParams.put("delaunaycut.seed", 1);
 
     std::array<Point3d, 8> hexah;
@@ -85,29 +85,31 @@ BOOST_AUTO_TEST_CASE(fuseCut_delaunayGraphCut)
 
     const std::string tempDirPath = std::filesystem::temp_directory_path().generic_string();
 
-    DelaunayGraphCut delaunayGC(mp);
+    
     ALICEVISION_LOG_TRACE("Creating dense point cloud witout support pts.");
 
     // delaunayGC.createDensePointCloud(&hexah[0], cams, &sfmData, nullptr);
     const float minDist = (hexah[0] - hexah[1]).size() / 1000.0f;
     // add points for cam centers
-    delaunayGC.addPointsFromCameraCenters(cams, minDist);
+    
+    fuseCut::PointCloud pointcloud(mp);
+    pointcloud.addPointsFromCameraCenters(cams, minDist);
     // add points from sfm
-    delaunayGC.addPointsFromSfM(&hexah[0], cams, sfmData);
+    pointcloud.addPointsFromSfM(&hexah[0], cams, sfmData);
+
+    fuseCut::Tetrahedralization tetrahedralization(pointcloud.getVertices());
+    fuseCut::GraphFiller gfiller(mp, pointcloud, tetrahedralization);
+    gfiller.build(cams);
+    gfiller.binarize();
 
     ALICEVISION_LOG_TRACE("Generated pts:");
-    for (size_t i = 0; i < delaunayGC._verticesCoords.size(); i++)
-        ALICEVISION_LOG_TRACE("[" << i << "]: " << delaunayGC._verticesCoords[i].x << ", " << delaunayGC._verticesCoords[i].y << ", "
-                                  << delaunayGC._verticesCoords[i].z);
-
-    delaunayGC.createGraphCut(&hexah[0], cams, tempDirPath + "/", tempDirPath + "/SpaceCamsTracks/", false, false);
-    /*
-    delaunayGC.computeDelaunay();
-    delaunayGC.displayStatistics();
-    delaunayGC.computeVerticesSegSize(true, 0.0f);
-    delaunayGC.voteFullEmptyScore(cams, tempDirPath);
-    delaunayGC.reconstructGC(&hexah[0]);
-    */
+    for (size_t i = 0; i < pointcloud.getVertices().size(); i++)
+    {
+        ALICEVISION_LOG_TRACE("[" << i << "]: " 
+                << pointcloud.getVertices()[i].x << ", " 
+                << pointcloud.getVertices()[i].y << ", "
+                << pointcloud.getVertices()[i].z);
+    }
 
     ALICEVISION_LOG_TRACE("CreateGraphCut Done.");
 }
