+++ src/aliceVision/fuseCut/Fuser.cpp
@@ -618,237 +618,6 @@ void Fuser::divideSpaceFromSfM(const sfmData::SfMData& sfmData, Point3d* hexah, 
     ALICEVISION_LOG_INFO("Estimate space done.");
 }
 
-Voxel Fuser::estimateDimensions(Point3d* vox, Point3d* newSpace, int scale, int maxOcTreeDim, const sfmData::SfMData* sfmData)
-{
-    const Point3d O = (vox[0] + vox[1] + vox[2] + vox[3] + vox[4] + vox[5] + vox[6] + vox[7]) / 8.0f;
-    Point3d vx = vox[1] - vox[0];
-    Point3d vy = vox[3] - vox[0];
-    Point3d vz = vox[4] - vox[0];
-    const float svx = vx.size();
-    const float svy = vy.size();
-    const float svz = vz.size();
-    vx = vx.normalize();
-    vy = vy.normalize();
-    vz = vz.normalize();
-
-    const float pointToJoinPixSizeDist = (float)_mp.userParams.get<double>("Fuser.pointToJoinPixSizeDist", 2.0f);
-    ALICEVISION_LOG_INFO("pointToJoinPixSizeDist: " << pointToJoinPixSizeDist);
-
-    float aAvPixelSize;
-
-    if (sfmData == nullptr)
-    {
-        // WARNING perf: reload all depth maps to compute the minPixelSize (minPixelSize consider only points in the hexahedron)
-        // Average 3D size for each pixel from all 3D points in the current voxel
-        const int maxPts = 1000000;
-        const int nAllPts = computeNumberOfAllPoints(_mp, scale);
-        const int stepPts = nAllPts / maxPts + 1;
-        aAvPixelSize = computeAveragePixelSizeInHexahedron(vox, stepPts, scale) * (float)std::max(scale, 1) * pointToJoinPixSizeDist;
-    }
-    else
-    {
-        aAvPixelSize = computeAveragePixelSizeInHexahedron(vox, *sfmData) * pointToJoinPixSizeDist;
-    }
-
-    Voxel maxDim;
-    maxDim.x = (int)ceil(svx / (aAvPixelSize * (float)maxOcTreeDim));
-    maxDim.y = (int)ceil(svy / (aAvPixelSize * (float)maxOcTreeDim));
-    maxDim.z = (int)ceil(svz / (aAvPixelSize * (float)maxOcTreeDim));
-
-    const Point3d vvx = vx * ((float)maxDim.x * ((aAvPixelSize * (float)maxOcTreeDim)));
-    const Point3d vvy = vy * ((float)maxDim.y * ((aAvPixelSize * (float)maxOcTreeDim)));
-    const Point3d vvz = vz * ((float)maxDim.z * ((aAvPixelSize * (float)maxOcTreeDim)));
-
-    newSpace[0] = O - vvx / 2.0f - vvy / 2.0f - vvz / 2.0f;
-    newSpace[1] = O + vvx - vvx / 2.0f - vvy / 2.0f - vvz / 2.0f;
-    newSpace[2] = O + vvx + vvy - vvx / 2.0f - vvy / 2.0f - vvz / 2.0f;
-    newSpace[3] = O + vvy - vvx / 2.0f - vvy / 2.0f - vvz / 2.0f;
-    newSpace[4] = O + vvz - vvx / 2.0f - vvy / 2.0f - vvz / 2.0f;
-    newSpace[5] = O + vvz + vvx - vvx / 2.0f - vvy / 2.0f - vvz / 2.0f;
-    newSpace[6] = O + vvz + vvx + vvy - vvx / 2.0f - vvy / 2.0f - vvz / 2.0f;
-    newSpace[7] = O + vvz + vvy - vvx / 2.0f - vvy / 2.0f - vvz / 2.0f;
-
-    ALICEVISION_LOG_INFO("Estimated: " << mvsUtils::num2str(maxDim.x) << " " << mvsUtils::num2str(maxDim.y) << " " << mvsUtils::num2str(maxDim.z));
-
-    ALICEVISION_LOG_DEBUG("optimal detail: " << static_cast<int>((vvx.size() / (float)maxDim.x) / aAvPixelSize) << " "
-                                             << static_cast<int>((vvy.size() / (float)maxDim.y) / aAvPixelSize) << " "
-                                             << static_cast<int>((vvz.size() / (float)maxDim.z) / aAvPixelSize));
-
-    return maxDim;
-}
-
-std::string generateTempPtsSimsFiles(const std::string& tmpDir,
-                                     mvsUtils::MultiViewParams& mp,
-                                     bool addRandomNoise,
-                                     float percNoisePts,
-                                     int noisPixSizeDistHalfThr)
-{
-    ALICEVISION_LOG_INFO("generating temp files.");
-    std::string depthMapsPtsSimsTmpDir = tmpDir + "depthMapsPtsSimsTmp/";
-
-    if (!fs::is_directory(depthMapsPtsSimsTmpDir))
-    {
-        fs::create_directory(depthMapsPtsSimsTmpDir);
-
-        const int scaleuse = 1;
-
-        StaticVector<Point2d>* minMaxDepths = new StaticVector<Point2d>();
-        minMaxDepths->reserve(mp.ncams);
-        minMaxDepths->resize_with(mp.ncams, Point2d(-1.0, -1.0));
-
-#pragma omp parallel for
-        for (int rc = 0; rc < mp.ncams; rc++)
-        {
-            int w = mp.getWidth(rc) / scaleuse;
-            int h = mp.getHeight(rc) / scaleuse;
-
-            StaticVector<Point3d>* pts = new StaticVector<Point3d>();
-            StaticVector<float>* sims = new StaticVector<float>();
-
-            pts->reserve(w * h);
-            sims->reserve(w * h);
-
-            image::Image<float> depthMap;
-            image::Image<float> simMap;
-
-            mvsUtils::readMap(rc, mp, mvsUtils::EFileType::depthMapFiltered, depthMap);
-            mvsUtils::readMap(rc, mp, mvsUtils::EFileType::simMapFiltered, simMap);
-
-            if (depthMap.size() != (w * h) || simMap.size() != (w * h))
-            {
-                throw std::runtime_error("Invalid image size");
-            }
-
-            if (addRandomNoise)
-            {
-                StaticVector<int>* idsAlive = new StaticVector<int>();
-                idsAlive->reserve(w * h);
-                for (int i = 0; i < w * h; i++)
-                {
-                    if (depthMap(i) > 0.0f)
-                    {
-                        idsAlive->push_back(i);
-                    }
-                }
-
-                int nnoisePts = ((percNoisePts / 100.0f) * (float)(idsAlive->size()));
-                const std::vector<int> randIdsAlive = mvsUtils::createRandomArrayOfIntegers(idsAlive->size());
-
-                srand(time(nullptr));
-
-                long t1 = clock();
-                for (int y = 0; y < h; ++y)
-                {
-                    for (int x = 0; x < w; ++x)
-                    {
-                        int id = y * w + x;
-                        int i = (*idsAlive)[randIdsAlive[id]];
-                        double depth = depthMap(i);
-
-                        double sim = simMap(i);
-                        if (depth > 0.0f)
-                        {
-                            Point3d p = mp.CArr[rc] +
-                                        (mp.iCamArr[rc] * Point2d((double)x * (double)scaleuse, (double)y * (double)scaleuse)).normalize() * depth;
-
-                            if (id < nnoisePts)
-                            {
-                                double pixSize = mp.getCamPixelSize(p, rc);
-                                int rid = rand() % (2 * noisPixSizeDistHalfThr + 1);
-                                rid = rid - noisPixSizeDistHalfThr;
-                                double rdepthAdd = pixSize * (double)rid;
-                                depth = depth + rdepthAdd;
-                                p = mp.CArr[rc] +
-                                    (mp.iCamArr[rc] * Point2d((double)x * (double)scaleuse, (double)y * (double)scaleuse)).normalize() * depth;
-                            }
-
-                            pts->push_back(p);
-                            sims->push_back(sim);
-                            if ((*minMaxDepths)[rc].x < 0.0f)
-                            {
-                                (*minMaxDepths)[rc].x = depth;
-                            }
-                            else
-                            {
-                                (*minMaxDepths)[rc].x = std::min((*minMaxDepths)[rc].x, depth);
-                            }
-                            if ((*minMaxDepths)[rc].y < 0.0f)
-                            {
-                                (*minMaxDepths)[rc].y = depth;
-                            }
-                            else
-                            {
-                                (*minMaxDepths)[rc].y = std::max((*minMaxDepths)[rc].y, depth);
-                            }
-                        }
-                    }
-                }
-                mvsUtils::printfElapsedTime(t1);
-
-                delete idsAlive;
-            }
-            else
-            {
-                long t1 = clock();
-                for (int x = 0; x < w; x++)
-                {
-                    for (int y = 0; y < h; y++)
-                    {
-                        int i = x * h + y;
-                        double depth = depthMap(i);
-                        double sim = simMap(i);
-                        if (depth > 0.0f)
-                        {
-                            Point3d p = mp.CArr[rc] +
-                                        (mp.iCamArr[rc] * Point2d((double)x * (double)scaleuse, (double)y * (double)scaleuse)).normalize() * depth;
-                            pts->push_back(p);
-                            sims->push_back(sim);
-                            if ((*minMaxDepths)[rc].x < 0.0f)
-                            {
-                                (*minMaxDepths)[rc].x = depth;
-                            }
-                            else
-                            {
-                                (*minMaxDepths)[rc].x = std::min((*minMaxDepths)[rc].x, depth);
-                            }
-                            if ((*minMaxDepths)[rc].y < 0.0f)
-                            {
-                                (*minMaxDepths)[rc].y = depth;
-                            }
-                            else
-                            {
-                                (*minMaxDepths)[rc].y = std::max((*minMaxDepths)[rc].y, depth);
-                            }
-                        }
-                    }
-                }
-                mvsUtils::printfElapsedTime(t1);
-            }
-
-            saveArrayToFile<Point3d>(depthMapsPtsSimsTmpDir + std::to_string(mp.getViewId(rc)) + "pts.bin", pts);
-            saveArrayToFile<float>(depthMapsPtsSimsTmpDir + std::to_string(mp.getViewId(rc)) + "sims.bin", sims);
-            delete pts;
-            delete sims;
-        }
-
-        saveArrayToFile<Point2d>(depthMapsPtsSimsTmpDir + "minMaxDepths.bin", minMaxDepths);
-        delete minMaxDepths;
-    }
-
-    return depthMapsPtsSimsTmpDir;
-}
-
-void deleteTempPtsSimsFiles(mvsUtils::MultiViewParams& mp, const std::string& depthMapsPtsSimsTmpDir)
-{
-    for (int rc = 0; rc < mp.ncams; rc++)
-    {
-        std::string ptsfn = depthMapsPtsSimsTmpDir + std::to_string(mp.getViewId(rc)) + "pts.bin";
-        std::string simsfn = depthMapsPtsSimsTmpDir + std::to_string(mp.getViewId(rc)) + "sims.bin";
-        remove(ptsfn.c_str());
-        remove(simsfn.c_str());
-    }
-    fs::remove_all(depthMapsPtsSimsTmpDir);
-}
 
 }  // namespace fuseCut
 }  // namespace aliceVision
